+++ src/utils.ts
@@ -1,4 +1,5 @@
 import {window} from 'vscode';
+import App from './app';
 
 export default class Utils
 {
@@ -89,5 +90,56 @@ export default class Utils
     public multiplyString(element: string, cnt: number, separator: string = ''): string
     {
         return this.fillArray(element, cnt).join(separator);
+    }
+
+    /**
+     * @param {object} obj
+     * @param {string} key
+     * @returns {key is keyof typeof obj}
+     */
+    public hasKey(obj: object, key: string): key is keyof typeof obj {
+        return key in obj;
+    }
+
+    /**
+     * @param {string} fullPath
+     * @returns {string}
+     */
+    public pathToNamespace(fullPath: string): string {
+        const autoloadData = App.instance.composer('autoload');
+        const relativePath = fullPath.replace(App.instance.workplacePath, '').replace('.php', '').substring(1);
+        let result = relativePath;
+        Object.keys(autoloadData).forEach((namespaceStart) => {
+            const searchString = autoloadData[namespaceStart];
+            if (relativePath.startsWith(searchString)) {
+                result = relativePath.replace(searchString, namespaceStart).replaceAll('/', "\\");
+            }
+        });
+
+        return result;
+    }
+
+    /**
+     * @param {string} fullPath
+     * @returns {[string, string]}
+     */
+    public splitPath(fullPath: string): [string, string] {
+        const lastSlashIndex = fullPath.lastIndexOf('/');
+        if (lastSlashIndex === -1) {
+            return ['', fullPath];
+        }
+
+        return [fullPath.substring(0, lastSlashIndex), fullPath.substring(lastSlashIndex + 1)];
+    }
+
+    /**
+     * @param {string} input
+     * @returns {string}
+     */
+    public capitalizeFirstCharTrimmed(input: string): string {
+        const trimmedInput = input.trim();
+        if (!trimmedInput) return trimmedInput;
+
+        return trimmedInput.charAt(0).toUpperCase() + trimmedInput.slice(1);
     }
 }